chore: Clean up old code from codebase. Moved old main to legacy_main.py. Removed garpos.py (#253)
* chore: Clean up main and move to legacy_main for now

* chore: Remove garpos and move constants to file
