--- src/gnatss/constants/garpos.py
@@ -1,9 +1,0 @@
-"""garpos.py
-
-Module for storing garpos constants
-"""
-
-ST = "T_transmit"  # Transmission time of acoustic signal
-RT = "T_receive"  # Reception time of acoustic signal
-MT = "MT_ID"  # ID of mirror transponder (should be consistent with Site-parameter file)
-TT = "TravelTime"  # Observed travel time
